core: extract independent LHR types
