+++ lighthouse-viewer/tsconfig.json
@@ -1,16 +1,12 @@
 {
   "extends": "../tsconfig",
   "compilerOptions": {
-    "typeRoots": [
-      "../node_modules/@types",
-      "../types",
-      "./types",
-    ],
     "esModuleInterop": true,
+    "composite": false,
   },
   "include": [
     "app/src/**/*.js",
-    "../types/*.d.ts",
+    "../types/**/*.d.ts",
     "./types/*.d.ts",
   ],
 }
