+++ types/lhr/i18n.d.ts
@@ -4,9 +4,8 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import {Util} from '../report/renderer/util.js';
-
-export type Locale = 'en-US'|'en'|'en-AU'|'en-GB'|'en-IE'|'en-SG'|'en-ZA'|'en-IN'|'ar-XB'|'ar'|'bg'|'ca'|'cs'|'da'|'de'|'el'|'en-XA'|'en-XL'|'es'|'es-419'|'es-AR'|'es-BO'|'es-BR'|'es-BZ'|'es-CL'|'es-CO'|'es-CR'|'es-CU'|'es-DO'|'es-EC'|'es-GT'|'es-HN'|'es-MX'|'es-NI'|'es-PA'|'es-PE'|'es-PR'|'es-PY'|'es-SV'|'es-US'|'es-UY'|'es-VE'|'fi'|'fil'|'fr'|'he'|'hi'|'hr'|'hu'|'gsw'|'id'|'in'|'it'|'iw'|'ja'|'ko'|'lt'|'lv'|'mo'|'nl'|'nb'|'no'|'pl'|'pt'|'pt-PT'|'ro'|'ru'|'sk'|'sl'|'sr'|'sr-Latn'|'sv'|'ta'|'te'|'th'|'tl'|'tr'|'uk'|'vi'|'zh'|'zh-HK'|'zh-TW';
+// TODO(bckenny): i18n types should not be necessary for the LHR, but too difficult
+// to unravel from audit-details right now.
 
 export type IcuMessage = {
   // NOTE: `i18nId` rather than just `id` to make tsc typing easier (vs type branding which won't survive JSON roundtrip).
@@ -50,25 +49,3 @@ export type FormattedIcu<T> = T extends IcuMessage ?
   Exclude<T, IcuMessage> | string :
   // Otherwise recurse into any properties and repeat.
   {[K in keyof T]: FormattedIcu<T[K]>};
-
-/**
- * Info about an `LH.IcuMessage` value that was localized to a string. Value
- * is either a
- *  - path (`_.set()` style) into the original object where the message was replaced
- *  - those paths and a set of values that were inserted into the localized strings.
- */
-type IcuMessagePath = string | {path: string, values: Record<string, string | number>};
-
-/**
- * A representation of `LH.IcuMessage`s that were in an object (e.g. `LH.Result`)
- * and have been replaced by localized strings. `LH.IcuMessagePaths` provides a
- * mapping that can be used to change the locale of the object again.
- * Keyed by ids from the locale message json files, values are information on
- * how to replace the string if needed.
- */
-export interface IcuMessagePaths {
-  [i18nId: string]: IcuMessagePath[];
-}
-
-/** Localized strings for the html report renderer. */
-export type I18NRendererStrings = typeof Util['UIStrings'];
