+++ types/lhr/treemap.d.ts
@@ -4,8 +4,8 @@
  * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
  */
 
-import Audit from './audit';
-import {Locale} from './i18n';
+import {Result as AuditResult} from './audit-result';
+import {Locale} from './settings';
 
 declare module Treemap {
   interface Options {
@@ -13,7 +13,7 @@ declare module Treemap {
       requestedUrl: string;
       finalUrl: string;
       audits: {
-        'script-treemap-data': Audit.Result;
+        'script-treemap-data': AuditResult;
       };
       configSettings: {
         locale: Locale;
