+++ types/lhr/tsconfig.json
@@ -1,0 +1,24 @@
+{
+  "compilerOptions": {
+    "module": "esnext",
+    "target": "esnext",
+    "esModuleInterop": true,
+    "moduleResolution": "node",
+
+    "strict": true,
+    "extendedDiagnostics": true,
+
+    // We only need the base JS definitions, no libs in node_modules, no DOM, etc.
+    "lib": ["esnext"],
+    "types": [],
+
+    // "listFiles": true,
+    // "noErrorTruncation": true,
+
+    "composite": true,
+    "outDir": "../../.tmp/tsbuildinfo/lhr",
+  },
+  "include": [
+    "*.d.ts",
+  ],
+}
